add component

+++ backend/src/logic/QuestionService.ts
@@ -128,14 +128,14 @@ export class QuestionService {
     const question = await this.questionAccess.findOneOrFail({ where: { id } });
 
     // delete & re-upload images
-    if (question.hasImage) {
+    if (data.image && data.image.length > 0 && question.hasImage) {
       const s3Objects = await this.awsUtil.listS3Objects(question.id);
       if (s3Objects.Contents)
         await this.awsUtil.deleteS3Objects(
           s3Objects.Contents.map((v) => v.Key ?? '').filter((v) => v !== '')
         );
     }
-    if (data.image) {
+    if (data.image && data.image.length > 0) {
       let n = 0;
       for (const i of data.image) {
         n = n + 1;
@@ -151,7 +151,7 @@ export class QuestionService {
     question.tags = thisTags ?? question.tags;
     question.youtube = data.youtube ?? question.youtube;
     question.hasSolution = data.hasSolution ?? question.hasSolution;
-    question.hasImage = data.image === undefined ? false : true;
+    question.hasImage = data.image === undefined ? question.hasImage : true;
 
     const updatedQuestion = await this.questionAccess.save(question);
